hospital_management_system/requirements.txt
Django==4.2.0
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7
python-dateutil==2.8.2
pytz==2023.3
google-generativeai==0.8.3
reportlab==4.0.7
stripe==7.8.0
razorpay==1.4.2
